Optimise metrics evaluation on manifest entries in `DataPlan.plan_files`
The `lambda` returned from `DataPlan._build_metrics_evaluator` can
be called a lot when there are many manifest entries. Currently, it
initialises a new `_InclusiveMetricsEvaluator` on every call to achieve
thread safety (it is called from the threads of the `ExecutorFactory`
thread pool). This is not particularly fast, with most of the time spent
inside `pydantic`, per profiling results.

As `_InclusiveMetricsEvaluator.eval` only mutates the count and bounds
instance variables (shallow mutation), a shallow copy can be used for
each thread, avoiding repeated expensive initialisation.

This change:
* Adds a thread-safe `eval_thread_safe` method to
  `_InclusiveMetricsEvaluator` that calls the `eval` method on a shallow
  copy
* Adds thread-safe (and cache stampede safe) lazy initialisation of the
  `_InclusiveMetricsEvaluator` used as a template for the shallow
  copies.
  * This uses functionality added in `cachetools` `6.x.y`, and the
    min version dependency pin is updated accordingly.
* Removes redundant calls to `_build_metrics_evaluator` in
  `DataPlan.plan_files`.